refactor(slack): split non-listener sections out of bot.ts
bot.ts had grown to 1759 lines mixing Slack API wrappers, live event
listeners, and several fully self-contained concerns that had nothing to
do with listening: Home tab rendering, message logging, channel history
backfill, and user/channel directory sync. This follows the split
convention already established by this directory's sibling files
(blocks.ts, session.ts, context.ts, response-lifecycle.ts) — pure/factory
functions taking explicit deps, imported back into bot.ts as the "center"
class.

- home-view.ts: buildHomeView(), pure function over injected deps
- message-logging.ts: logUserMessage()/logExternalMessagingBotMessage()
  plus their buildSlackAppMessageText()/collectSlackText() helpers
- backfill.ts: backfillChannel()/backfillAllChannels(), keeps
  backfillChannel exported since test/slack-bot.test.ts calls it directly
  via (bot as any) — bot.ts keeps a same-signature private wrapper so that
  test needed zero changes
- user-channel-sync.ts: fetchUsers()/fetchChannels(), mutating the passed
  in Maps in place

Deliberately out of scope: the actual Bolt.js listener bodies
(handleAppMention, handleMessageEvent, handleSlashCommand,
handleBlockAction) stay in bot.ts — they're tightly coupled to live
message-handling state and warrant a separate, more careful pass.

bot.ts: 1759 -> 1341 lines. No public API change; all 668 tests pass
unchanged, including slack-bot.test.ts's private-method casts.
