+++ src/adapters/slack/bot.ts
@@ -1,8 +1,7 @@
 import { SocketModeClient } from "@slack/socket-mode";
 import type { KnownBlock } from "@slack/types";
 import { WebClient } from "@slack/web-api";
-import { existsSync, readFileSync } from "fs";
-import { readFile } from "fs/promises";
+import { readFileSync } from "fs";
 import { basename, join } from "path";
 import type {
   MessagingBot,
@@ -49,6 +48,14 @@ import {
 } from "./session.js";
 import { reportUserFacingError } from "../../observability/sentry.js";
 import { renderSlackBlocks } from "./blocks.js";
+import { backfillAllChannels, backfillChannel as backfillChannelImpl } from "./backfill.js";
+import { buildHomeView } from "./home-view.js";
+import {
+  logExternalMessagingBotMessage as logExternalMessagingBotMessageImpl,
+  logUserMessage as logUserMessageImpl,
+  type ExternalBotMessageEvent,
+} from "./message-logging.js";
+import { fetchChannels, fetchUsers } from "./user-channel-sync.js";
 
 const SLACK_EVENT_ANCHOR_TEXT = "Working on it...";
 
@@ -62,41 +69,6 @@ function slackIsRateLimited(err: Error): boolean {
 
 const slackRetry = <T>(fn: () => Promise<T>): Promise<T> =>
   withRetry(fn, { isRateLimited: slackIsRateLimited });
-
-function collectSlackText(value: unknown, parts: string[]): void {
-  if (value === null || value === undefined) return;
-  if (typeof value === "string") {
-    const trimmed = value.trim();
-    if (trimmed) parts.push(trimmed);
-    return;
-  }
-  if (Array.isArray(value)) {
-    for (const item of value) collectSlackText(item, parts);
-    return;
-  }
-  if (typeof value !== "object") return;
-
-  const obj = value as Record<string, unknown>;
-  for (const key of ["text", "fallback", "title", "value"] as const) {
-    collectSlackText(obj[key], parts);
-  }
-  collectSlackText(obj.fields, parts);
-  collectSlackText(obj.elements, parts);
-  collectSlackText(obj.blocks, parts);
-}
-
-function buildSlackAppMessageText(event: {
-  text?: string;
-  blocks?: unknown[];
-  attachments?: unknown[];
-}): string {
-  const parts: string[] = [];
-  collectSlackText(event.text, parts);
-  collectSlackText(event.blocks, parts);
-  collectSlackText(event.attachments, parts);
-  const deduped = parts.filter((part, index) => parts.indexOf(part) === index);
-  return deduped.join("\n");
-}
 
 // ---------------------------------------------------------------------------
 // Shared mrkdwn truncation helper
@@ -181,7 +153,10 @@ export class SlackMessagingBot implements MessagingBot {
     this.botId = typeof auth.bot_id === "string" ? auth.bot_id : null;
     this.teamId = typeof auth.team_id === "string" ? auth.team_id : null;
 
-    await Promise.all([this.fetchUsers(), this.fetchChannels()]);
+    await Promise.all([
+      fetchUsers({ webClient: this.webClient, users: this.users }),
+      fetchChannels({ webClient: this.webClient, users: this.users, channels: this.channels }),
+    ]);
     log.logInfo(`Loaded ${this.channels.size} channels, ${this.users.size} users`);
 
     // Record startup time before opening the socket. Slack may replay older events;
@@ -194,7 +169,10 @@ export class SlackMessagingBot implements MessagingBot {
 
     log.logConnected("Slack");
 
-    void this.backfillAllChannels(this.startupTs).catch((error) => {
+    void backfillAllChannels(
+      { ...this.backfillDeps(), channels: this.channels },
+      this.startupTs,
+    ).catch((error) => {
       log.logWarning("Slack backfill failed", String(error));
     });
   }
@@ -583,159 +561,11 @@ export class SlackMessagingBot implements MessagingBot {
   }
 
   private buildHomeView(): { type: "home"; blocks: KnownBlock[] } {
-    const blocks: object[] = [
-      {
-        type: "section",
-        text: {
-          type: "mrkdwn",
-          text: `*${PRODUCT_NAME}*\nStart a new task or check on running work.`,
-        },
-        accessory: {
-          type: "image",
-          image_url: "https://media1.tenor.com/m/lfDATg4Bhc0AAAAC/happy-cat.gif",
-          alt_text: PRODUCT_NAME,
-        },
-      },
-    ];
-
-    // --- Running tasks ---
-    const runningSessions = this.handler.getRunningSessions();
-
-    blocks.push(
-      { type: "divider" },
-      {
-        type: "header",
-        text: {
-          type: "plain_text",
-          text: `Running Tasks (${runningSessions.length})`,
-          emoji: true,
-        },
-      },
-    );
-
-    if (runningSessions.length === 0) {
-      blocks.push({
-        type: "context",
-        elements: [{ type: "mrkdwn", text: "_No tasks running right now._" }],
-      });
-    } else {
-      // Threshold for "stuck" detection (10 minutes)
-      const STUCK_THRESHOLD_MS = 10 * 60 * 1000;
-
-      for (const session of runningSessions) {
-        const channelId = session.sessionKey.split(":")[0];
-        const channel = this.channels.get(channelId);
-        const channelName = channel ? `#${channel.name}` : channelId;
-        const elapsed = Math.floor((Date.now() - session.startedAt) / 60000);
-        const elapsedStr = elapsed < 1 ? "<1 min" : `${elapsed} min`;
-
-        // Check if task might be stuck
-        const lastActivity = session.lastActivityAt ? Date.now() - session.lastActivityAt : 0;
-        const isStuck = lastActivity > STUCK_THRESHOLD_MS;
-        const statusText = isStuck ? "_stuck_" : "_running_";
-
-        // Build status line: channel · status · time · step
-        let statusLine = `${statusText} · ${elapsedStr}`;
-        if (session.currentTool) {
-          statusLine += ` · ${session.currentTool}`;
-        }
-        if (isStuck && lastActivity > 0) {
-          const inactiveMin = Math.floor(lastActivity / 60000);
-          statusLine += ` · idle ${inactiveMin}m`;
-        }
-
-        // Use context block for gray small text (like "No scheduled jobs.")
-        blocks.push({
-          type: "context",
-          elements: [
-            {
-              type: "mrkdwn",
-              text: `*${channelName}* · ${statusLine}`,
-            },
-          ],
-        });
-
-        // Add Force Stop button as separate element if stuck
-        if (isStuck) {
-          blocks.push({
-            type: "context",
-            elements: [
-              {
-                type: "mrkdwn",
-                text: " ",
-              },
-              {
-                type: "button",
-                text: { type: "plain_text", text: "Force Stop", emoji: true },
-                action_id: `force_stop_${session.sessionKey.replace(/:/g, "_")}`,
-                style: "danger",
-              },
-            ],
-          });
-        }
-      }
-    }
-
-    // --- Cron jobs ---
-    const periodicEvents = this.eventsWatcher?.getPeriodicEvents() ?? [];
-
-    blocks.push(
-      { type: "divider" },
-      {
-        type: "header",
-        text: {
-          type: "plain_text",
-          text: `Scheduled Jobs (${periodicEvents.length})`,
-          emoji: true,
-        },
-      },
-    );
-
-    if (periodicEvents.length === 0) {
-      blocks.push({
-        type: "context",
-        elements: [{ type: "mrkdwn", text: "_No scheduled jobs._" }],
-      });
-    } else {
-      for (const ev of periodicEvents) {
-        const channelLabel =
-          ev.platform === "slack"
-            ? (() => {
-                const channel = this.channels.get(ev.conversationId);
-                const channelName = channel ? `#${channel.name}` : ev.conversationId;
-                return `${ev.platform}:${channelName}`;
-              })()
-            : `${ev.platform}:${ev.conversationId}`;
-        const nextStr = ev.nextRun
-          ? new Date(ev.nextRun).toLocaleString("en-US", {
-              month: "short",
-              day: "numeric",
-              hour: "2-digit",
-              minute: "2-digit",
-            })
-          : "—";
-        blocks.push({
-          type: "section",
-          text: {
-            type: "mrkdwn",
-            text: `*${ev.text}*\n└ \`${ev.schedule}\` · ${channelLabel} · Next: ${nextStr}`,
-          },
-        });
-      }
-    }
-
-    // --- Footer ---
-    blocks.push(
-      { type: "divider" },
-      {
-        type: "context",
-        elements: [
-          { type: "mrkdwn", text: "💡 @mention in a channel or send a DM to start a new task" },
-        ],
-      },
-    );
-
-    return { type: "home", blocks: blocks as KnownBlock[] };
+    return buildHomeView({
+      getRunningSessions: () => this.handler.getRunningSessions(),
+      channels: this.channels,
+      getPeriodicEvents: () => this.eventsWatcher?.getPeriodicEvents() ?? [],
+    });
   }
 
   private resolveStopTarget(channelId: string, threadTs?: string): string | null {
@@ -1455,305 +1285,57 @@ export class SlackMessagingBot implements MessagingBot {
   /**
    * Log a user message to log.jsonl after attachments are ready.
    */
-  private async logUserMessage(event: SlackEvent): Promise<Attachment[]> {
-    const user = this.users.get(event.user);
-    let attachments: Attachment[] = [];
-    let attachmentError: unknown;
-    if (event.files) {
-      try {
-        attachments = await this.store.processAttachments(event.channel, event.files, event.ts);
-      } catch (err) {
-        attachmentError = err;
-      }
-    }
-    // Always write the text log, even if attachment processing failed — we want
-    // a record of the user message regardless of file-handling errors.
-    this.logToFile(event.channel, {
-      date: new Date(parseFloat(event.ts) * 1000).toISOString(),
-      ts: event.ts,
-      threadTs: event.thread_ts,
-      user: event.user,
-      userName: user?.userName,
-      displayName: user?.displayName,
-      text: event.text,
-      attachments,
-      isMessagingBot: false,
-    });
-    if (attachmentError) throw attachmentError;
-    return attachments;
+  private logUserMessage(event: SlackEvent): Promise<Attachment[]> {
+    return logUserMessageImpl(
+      {
+        users: this.users,
+        processAttachments: this.store.processAttachments.bind(this.store),
+        logToFile: this.logToFile.bind(this),
+      },
+      event,
+    );
   }
 
-  private async logExternalMessagingBotMessage(event: {
-    channel: string;
-    ts: string;
-    thread_ts?: string;
-    text?: string;
-    subtype?: string;
-    bot_id?: string;
-    app_id?: string;
-    username?: string;
-    bot_profile?: { app_id?: string; name?: string; real_name?: string };
-    blocks?: unknown[];
-    attachments?: unknown[];
-    files?: Array<{ name: string; url_private_download?: string; url_private?: string }>;
-  }): Promise<Attachment[]> {
-    const attachments = event.files
-      ? await this.store.processAttachments(event.channel, event.files, event.ts)
-      : [];
-    const botName =
-      event.username ?? event.bot_profile?.name ?? event.bot_profile?.real_name ?? event.bot_id;
-    this.logToFile(event.channel, {
-      date: new Date(parseFloat(event.ts) * 1000).toISOString(),
-      ts: event.ts,
-      threadTs: event.thread_ts,
-      user: event.bot_id ? `bot:${event.bot_id}` : "external-bot",
-      userName: botName,
-      displayName: botName,
-      text: buildSlackAppMessageText(event),
-      attachments,
-      isMessagingBot: true,
-      botId: event.bot_id,
-      appId: event.app_id ?? event.bot_profile?.app_id,
-      subtype: event.subtype,
-    });
-    return attachments;
+  private logExternalMessagingBotMessage(event: ExternalBotMessageEvent): Promise<Attachment[]> {
+    return logExternalMessagingBotMessageImpl(
+      {
+        users: this.users,
+        processAttachments: this.store.processAttachments.bind(this.store),
+        logToFile: this.logToFile.bind(this),
+      },
+      event,
+    );
   }
 
   // ==========================================================================
   // Private - Backfill
   // ==========================================================================
 
-  private async getExistingTimestamps(channelId: string): Promise<Set<string>> {
-    const logPath = join(this.workingDir, channelId, "log.jsonl");
-    const timestamps = new Set<string>();
-    if (!existsSync(logPath)) return timestamps;
-
-    const content = await readFile(logPath, "utf-8");
-    const lines = content.trim().split("\n").filter(Boolean);
-    for (let i = 0; i < lines.length; i++) {
-      try {
-        const entry = JSON.parse(lines[i]);
-        if (entry.ts) timestamps.add(entry.ts);
-      } catch (err) {
-        log.logWarning(
-          `Skipping malformed log entry at ${logPath}:${i + 1}`,
-          err instanceof Error ? err.message : String(err),
-        );
-      }
-    }
-    return timestamps;
-  }
-
-  private async backfillChannel(channelId: string, upperBoundTs?: string): Promise<number> {
-    const existingTs = await this.getExistingTimestamps(channelId);
-
-    // Find the biggest ts in log.jsonl
-    let lastLoggedTs: string | undefined;
-    for (const ts of existingTs) {
-      if (!lastLoggedTs || parseFloat(ts) > parseFloat(lastLoggedTs)) lastLoggedTs = ts;
-    }
-
-    type Message = {
-      user?: string;
-      bot_id?: string;
-      app_id?: string;
-      username?: string;
-      bot_profile?: { app_id?: string; name?: string; real_name?: string };
-      blocks?: unknown[];
-      attachments?: unknown[];
-      text?: string;
-      ts?: string;
-      thread_ts?: string;
-      subtype?: string;
-      files?: Array<{ name: string }>;
+  private backfillDeps(): {
+    workingDir: string;
+    webClient: WebClient;
+    botUserId: string | null;
+    botId: string | null;
+    users: Map<string, SlackUser>;
+    stripOwnMention: (text?: string) => string;
+    processAttachments: ChannelStore["processAttachments"];
+    logToFile: (channel: string, entry: object) => void;
+    logExternalMessagingBotMessage: (event: ExternalBotMessageEvent) => Promise<Attachment[]>;
+  } {
+    return {
+      workingDir: this.workingDir,
+      webClient: this.webClient,
+      botUserId: this.botUserId,
+      botId: this.botId,
+      users: this.users,
+      stripOwnMention: (text?: string) => this.stripOwnMention(text),
+      processAttachments: this.store.processAttachments.bind(this.store),
+      logToFile: this.logToFile.bind(this),
+      logExternalMessagingBotMessage: (event) => this.logExternalMessagingBotMessage(event),
     };
-    const allMessages: Message[] = [];
-
-    let cursor: string | undefined;
-    let pageCount = 0;
-    const maxPages = 3;
-
-    do {
-      const result = await this.webClient.conversations.history({
-        channel: channelId,
-        oldest: lastLoggedTs, // Only fetch messages newer than what we have
-        latest: upperBoundTs, // Do not race live socket events after startup
-        inclusive: false,
-        limit: 1000,
-        cursor,
-      });
-      if (result.messages) {
-        allMessages.push(...(result.messages as Message[]));
-      }
-      cursor = result.response_metadata?.next_cursor;
-      pageCount++;
-    } while (cursor && pageCount < maxPages);
-
-    // Filter: include mikan's messages, external app/bot messages, and user messages.
-    const relevantMessages = allMessages.filter((msg) => {
-      if (!msg.ts || existingTs.has(msg.ts)) return false; // Skip duplicates
-      if (msg.user === this.botUserId) return true;
-      const isExternalMessagingBotMessage = !!msg.bot_id || msg.subtype === "bot_message";
-      if (isExternalMessagingBotMessage) {
-        if (this.botId && msg.bot_id === this.botId) return false;
-        if (
-          msg.subtype !== undefined &&
-          msg.subtype !== "bot_message" &&
-          msg.subtype !== "file_share"
-        ) {
-          return false;
-        }
-        return (
-          !!msg.text ||
-          !!(msg.files && msg.files.length > 0) ||
-          !!msg.blocks?.length ||
-          !!msg.attachments?.length
-        );
-      }
-      if (msg.subtype !== undefined && msg.subtype !== "file_share") return false;
-      if (!msg.user) return false;
-      if (!msg.text && (!msg.files || msg.files.length === 0)) return false;
-      return true;
-    });
-
-    // Reverse to chronological order
-    relevantMessages.reverse();
-
-    // Log each message to log.jsonl
-    for (const msg of relevantMessages) {
-      const isMikanMessage = msg.user === this.botUserId;
-      const isExternalMessagingBotMessage =
-        !isMikanMessage && (!!msg.bot_id || msg.subtype === "bot_message");
-      if (isExternalMessagingBotMessage) {
-        await this.logExternalMessagingBotMessage({ ...msg, channel: channelId, ts: msg.ts! });
-        continue;
-      }
-
-      const user = this.users.get(msg.user!);
-      const text = this.stripOwnMention(msg.text);
-      const attachments = msg.files
-        ? await this.store.processAttachments(channelId, msg.files, msg.ts!)
-        : [];
-
-      this.logToFile(channelId, {
-        date: new Date(parseFloat(msg.ts!) * 1000).toISOString(),
-        ts: msg.ts!,
-        threadTs: msg.thread_ts,
-        user: isMikanMessage ? "bot" : msg.user!,
-        userName: isMikanMessage ? undefined : user?.userName,
-        displayName: isMikanMessage ? undefined : user?.displayName,
-        text,
-        attachments,
-        isMessagingBot: isMikanMessage,
-      });
-    }
-
-    return relevantMessages.length;
   }
 
-  private async backfillAllChannels(upperBoundTs?: string): Promise<void> {
-    const startTime = Date.now();
-
-    // Only backfill channels that already have a log.jsonl (mikan has interacted with them before)
-    const channelsToBackfill: Array<[string, SlackChannel]> = [];
-    for (const [channelId, channel] of this.channels) {
-      const logPath = join(this.workingDir, channelId, "log.jsonl");
-      if (existsSync(logPath)) {
-        channelsToBackfill.push([channelId, channel]);
-      }
-    }
-
-    log.logBackfillStart(channelsToBackfill.length);
-
-    let totalMessages = 0;
-    for (const [channelId, channel] of channelsToBackfill) {
-      try {
-        const count = await this.backfillChannel(channelId, upperBoundTs);
-        if (count > 0) log.logBackfillChannel(channel.name, count);
-        totalMessages += count;
-      } catch (error) {
-        log.logWarning(`Failed to backfill #${channel.name}`, String(error));
-      }
-
-      // Add delay between channels to avoid hitting Slack rate limits
-      if (channelId !== channelsToBackfill[channelsToBackfill.length - 1][0]) {
-        await new Promise((resolve) => setTimeout(resolve, 500));
-      }
-    }
-
-    const durationMs = Date.now() - startTime;
-    log.logBackfillComplete(totalMessages, durationMs);
-  }
-
-  // ==========================================================================
-  // Private - Fetch Users/Channels
-  // ==========================================================================
-
-  private async fetchUsers(): Promise<void> {
-    let cursor: string | undefined;
-    do {
-      const result = await this.webClient.users.list({ limit: 200, cursor });
-      const members = result.members as
-        | Array<{ id?: string; name?: string; real_name?: string; deleted?: boolean }>
-        | undefined;
-      if (members) {
-        for (const u of members) {
-          if (u.id && u.name && !u.deleted) {
-            this.users.set(u.id, {
-              id: u.id,
-              userName: u.name,
-              displayName: u.real_name || u.name,
-            });
-          }
-        }
-      }
-      cursor = result.response_metadata?.next_cursor;
-    } while (cursor);
-  }
-
-  private async fetchChannels(): Promise<void> {
-    // Fetch public/private channels
-    let cursor: string | undefined;
-    do {
-      const result = await this.webClient.conversations.list({
-        types: "public_channel,private_channel",
-        exclude_archived: true,
-        limit: 200,
-        cursor,
-      });
-      const channels = result.channels as
-        | Array<{ id?: string; name?: string; is_member?: boolean }>
-        | undefined;
-      if (channels) {
-        for (const c of channels) {
-          if (c.id && c.name && c.is_member) {
-            this.channels.set(c.id, { id: c.id, name: c.name });
-          }
-        }
-      }
-      cursor = result.response_metadata?.next_cursor;
-    } while (cursor);
-
-    // Also fetch DM channels (IMs)
-    cursor = undefined;
-    do {
-      const result = await this.webClient.conversations.list({
-        types: "im",
-        limit: 200,
-        cursor,
-      });
-      const ims = result.channels as Array<{ id?: string; user?: string }> | undefined;
-      if (ims) {
-        for (const im of ims) {
-          if (im.id) {
-            // Use user's name as channel name for DMs
-            const user = im.user ? this.users.get(im.user) : undefined;
-            const name = user ? `DM:${user.userName}` : `DM:${im.id}`;
-            this.channels.set(im.id, { id: im.id, name });
-          }
-        }
-      }
-      cursor = result.response_metadata?.next_cursor;
-    } while (cursor);
+  private backfillChannel(channelId: string, upperBoundTs?: string): Promise<number> {
+    return backfillChannelImpl(this.backfillDeps(), channelId, upperBoundTs);
   }
 }
